wallet: move ImportPrivateKey to import.go
